// typechain/DeBridgeGate.d.ts
export type ClaimedEvent = TypedEvent<
  [string, string, BigNumber, string, BigNumber, BigNumber, string, boolean] & {
    submissionId: string;
    debridgeId: string;
    amount: BigNumber;
    receiver: string;
    nonce: BigNumber;
    chainIdFrom: BigNumber;
    autoParams: string;
    isNativeToken: boolean;
  }
>;
// ...
export type SentEvent = TypedEvent<
  [
    string,
    string,
    BigNumber,
    string,
    BigNumber,
    BigNumber,
    number,
    [BigNumber, BigNumber, BigNumber, boolean, boolean] & {
      receivedAmount: BigNumber;
      fixFee: BigNumber;
      transferFee: BigNumber;
      useAssetFee: boolean;
      isNativeToken: boolean;
    },
    string,
    string
  ] & {
    submissionId: string;
    debridgeId: string;
    amount: BigNumber;
    receiver: string;
    nonce: BigNumber;
    chainIdTo: BigNumber;
    referralCode: number;
    feeParams: [BigNumber, BigNumber, BigNumber, boolean, boolean] & {
      receivedAmount: BigNumber;
      fixFee: BigNumber;
      transferFee: BigNumber;
      useAssetFee: boolean;
      isNativeToken: boolean;
    };
    autoParams: string;
    nativeSender: string;
  }
>;